move provider definition up to fix #315
